refactor: update public api and prelude mod
